Add files via upload
back when i started getting into ARM systems and my first raspberry pi, i followed a youtube tutorial to make this. no copying and pasting, which really helped me slow down and think about what i was typing in rather than a quick paste job. also allowed me to fine-tune and weak small settings to make it how it liked it instead of a carbon copy of the one in the python video